[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 밤밤(홍우현) 미션 제출합니다. (#425)
* test(testConfig): 테스트용 헬퍼 분리

* refactor(application.properties): 타임존을 설정 파일로 분리

* docs(readme): 사이클 2 기능 목록 작성

* docs(readme): 예외 처리 응답 상태 코드 수정

* feat(exception): 중복 예약 검증 구현

* feat(reservationException): 예약 요청 검증 구현

* feat(reservationService): 날짜 검증 기능 구현

* feat(reservationService): 예약 시 시간 검증 기능 구현

* feat(reservationTimeRepository): 예약 시간 삭제 검증 기능 구현

* feat(exception): 유효하지 않은 입력값 검증 기능 구현

* feat(reservation): 본인 예약 취소 기능 구현

* feat(reservationController): 본인 예약 변경 api 구현

* refactor(service): 단순 조회에서 readonly 트랜젝션 제거

* refactor(repository): 데이터 저장 및 삭제 시 검증 책임을 서비스로 이동

* refactor(gloabalExceptionHandler): Exception 처리 핸들러를 RuntimeException 처리 핸들러로 변경

* test(repository): 이름 기반 조회 및 시간 기반 조회 테스트 추가

* refactor(reservationTimeService): 이용 가능 시간 조회에 대한 검증 로직 추가

* feat(themeController): 인기 테마 조회 시 쿼리 파라미터 검증 추가

* feat(reservationController): 이름 형식 검증 로직 추가

* refactor(controller): 관리자 컨트롤러 분리

* refactor(service): 메서드 순서 정리

* refactor(messageConverter): 커스텀 messageConverter 등록

* refactor(auth):인증/인가에 인터셉터를 사용하도록 변경

* refactor(global):패키지 구조 개선

* refactor(globalExceptionHandler): InternalServerError 응답 메시지 수정

* refactor(reservationService): 본인 예약 변경 시 예약 중복 검증 추가

* refactor(globalExceptionHandler): request 파라미터 누락 예외 핸들러 추가

* refactor(exception): 패키지 구조 수정

* refactor(globalExceptionHandler): MissingServletRequestParameterException 핸들러가 응답 메시지를 ErrorResponse로 감싸도록 수정

* test(reservationOwnerTest): 예약 변경 및 삭제에 대한 검증 테스트 추가

* test(exceptionTest): 응답 메시지 검증 추가

* refactor(service): 동시성 처리 구현

* refactor(service): 동시성 처리 로직 보강

* refactor(concurrencyTesst): 동시성 처리 테스트 보강

* refactor(repository): 메서드 순서 정렬

* refactor(exception): 커스텀 예외들을 상속 계층으로 재구성

* refactor(global): 패키지 구조 재배치

* test(repository, service): Unique 제약 조건에 대한 테스트 구체화

* refactor(reservationService): 삭제에서 만료 검증 로직 분리

* test(ExceptionTest): 잘못된 날짜 수정

* refactor(reservationController): Authorization 헤더 검증 삭제

* refactor(reservation): 이르 비교 로직 캡슐하

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -1,14 +1,17 @@
 package roomescape.theme.service;
 
 import java.util.List;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.theme.domain.Theme;
+import roomescape.theme.exception.DuplicateThemeException;
+import roomescape.theme.exception.ThemeInUseException;
+import roomescape.theme.exception.ThemeNotFoundException;
 import roomescape.theme.repository.ThemeRepository;
 import roomescape.theme.service.dto.ThemeCommand;
 
 @Service
-@Transactional(readOnly = true)
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
@@ -19,21 +22,44 @@ public class ThemeService {
 
     @Transactional
     public Theme registerTheme(ThemeCommand command) {
-        return themeRepository.save(
-                Theme.of(
-                        command.name(),
-                        command.description(),
-                        command.thumbnailUrl()
-                )
-        );
-    }
+        if (themeRepository.existByName(command.name())) {
+            throw new DuplicateThemeException();
+        }
 
-    @Transactional
-    public void removeThemeById(Long id) {
-        themeRepository.deleteById(id);
+
+        try {
+            return themeRepository.save(
+                    Theme.of(
+                            command.name(),
+                            command.description(),
+                            command.thumbnailUrl()
+                    )
+            );
+        } catch (DataIntegrityViolationException e) {
+            throw new DuplicateThemeException();
+        }
+
     }
 
     public List<Theme> findAllThemes() {
         return themeRepository.findAll();
+    }
+
+    @Transactional
+    public void removeThemeById(Long id) {
+        if (themeRepository.findById(id).isEmpty()) {
+            throw new ThemeNotFoundException();
+        }
+
+        try {
+            int affectedRow = themeRepository.deleteById(id);
+            int nonAffected = 0;
+
+            if (affectedRow == nonAffected) {
+                throw new ThemeNotFoundException();
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new ThemeInUseException();
+        }
     }
 }
